requirements.txt
django
feincms
sorl-thumbnail
django-admin-tools
django-autocomplete
django-compressor
django-debug-toolbar
django-mptt
django-require
